+++ src/validator/remp_manager.rs
@@ -18,10 +18,12 @@ use std::{
     sync::Arc,
     time::Duration
 };
+use std::cmp::{max, Reverse};
+use std::collections::BinaryHeap;
 
 use ton_block::{BlockIdExt, CatchainConfig, Message, ShardIdent, UnixTime32};
 use ton_api::ton::ton_node::RempMessageStatus;
-use ton_types::{error, KeyId, Result, SliceData, UInt256};
+use ton_types::{error, fail, KeyId, Result, SliceData, UInt256};
 
 use crate::{
     config::RempConfig,
@@ -35,8 +37,12 @@ use crate::{
 
 #[cfg(feature = "telemetry")]
 use std::time::Instant;
+use std::time::SystemTime;
 #[cfg(feature = "telemetry")]
 use adnl::telemetry::Metric;
+use chrono::{DateTime, Utc};
+use crossbeam_channel::TryRecvError;
+use rand::Rng;
 
 pub struct RempInterfaceQueues {
     message_cache: Arc<MessageCache>,
@@ -46,6 +52,109 @@ pub struct RempInterfaceQueues {
         crossbeam_channel::Sender<Arc<RmqMessage>>,
     pub response_receiver: 
         crossbeam_channel::Receiver<(UInt256, Arc<RmqMessage>, RempMessageStatus)>
+}
+
+pub struct RempDelayer {
+    max_incoming_broadcast_delay_millis: u32,
+    random_seed: u64,
+
+    pub incoming_receiver: crossbeam_channel::Receiver<Arc<RmqMessage>>,
+    pub delayed_incoming_sender: crossbeam_channel::Sender<Arc<RmqMessage>>,
+    delay_heap: MutexWrapper<(BinaryHeap<(Reverse<SystemTime>, UInt256)>, HashMap<UInt256, Arc<RmqMessage>>)>,
+}
+
+impl RempDelayer {
+    pub fn new (random_seed: u64, options: &RempConfig,
+                incoming_receiver: crossbeam_channel::Receiver<Arc<RmqMessage>>,
+                delayed_incoming_sender: crossbeam_channel::Sender<Arc<RmqMessage>>) -> Self {
+        Self {
+            max_incoming_broadcast_delay_millis: options.get_max_incoming_broadcast_delay_millis(),
+            random_seed,
+            incoming_receiver,
+            delayed_incoming_sender,
+            delay_heap: MutexWrapper::new((BinaryHeap::new(), HashMap::new()), "DelayerHeap".to_string()),
+        }
+    }
+
+    async fn push_delayer(&self, activation_time: SystemTime, msg: Arc<RmqMessage>) {
+        let tm: DateTime<Utc> = activation_time.into();
+        self.delay_heap.execute_sync(|(h,m)| {
+            if !m.contains_key(&msg.message_id) {
+                log::trace!(target: "remp", "Delaying REMP message {} till {}", msg, tm.format("%T"));
+                h.push((Reverse(activation_time), msg.message_id.clone()));
+                m.insert(msg.message_id.clone(), msg);
+            }
+            else {
+                log::trace!(target: "remp", "REMP message {} is already waiting", msg);
+            }
+        }).await
+    }
+
+    async fn pop_delayer(&self) -> Option<Arc<RmqMessage>> {
+        let now = SystemTime::now();
+        let res = self.delay_heap.execute_sync(|(h,m)| {
+            match h.peek() {
+                Some((Reverse(tm), _)) if tm > &now => return None,
+                None => return None,
+                _ => ()
+            };
+
+            match h.pop() {
+                Some((tm, id)) => m.remove(&id).map(|msg| (tm,msg)),
+                None => return None
+            }
+        }).await;
+
+        if let Some((tm,msg)) = res {
+            let tm: DateTime<Utc> = tm.0.into();
+            log::trace!(target: "remp", "Sending further REMP message {}, delayed till {}", msg, tm.format("%T"));
+            Some(msg)
+        }
+        else {
+            None
+        }
+    }
+
+    pub async fn len(&self) -> usize {
+        let (h,m) = self.delay_heap.execute_sync(|(h,m)| (h.len(),m.len())).await;
+        if h != m {
+            log::error!(target: "remp", "Different number of entries in Delay BinaryHeap and Delay HashMap: {} != {}", h, m);
+        }
+        max (h,m)
+    }
+
+    async fn poll_incoming_once(&self) -> Result<Option<Arc<RmqMessage>>> {
+        if let Some(msg) = self.pop_delayer().await {
+            return Ok(Some(msg))
+        }
+
+        loop {
+            match self.incoming_receiver.try_recv() {
+                Ok(msg) => {
+                    if msg.has_no_source_key() && self.max_incoming_broadcast_delay_millis > 0 {
+                        let delay = (msg.message_id.first_u64() + self.random_seed) % (self.max_incoming_broadcast_delay_millis as u64);
+                        self.push_delayer(SystemTime::now() + Duration::from_millis(delay), msg).await;
+                    }
+                    else {
+                        // self.push_delayer(SystemTime::now() + Duration::from_millis(100), msg).await; // TODO: debugging code
+                        return Ok(Some(msg));
+                    }
+                }
+                Err(TryRecvError::Empty) => return Ok(None),
+                Err(TryRecvError::Disconnected) => fail!("REMP incoming queue disconnected")
+            }
+        }
+    }
+
+    pub async fn poll_incoming(&self) -> Result<usize> {
+        let mut cnt = 0;
+        while let Some(msg) = self.poll_incoming_once().await? {
+            cnt += 1;
+            self.delayed_incoming_sender.send(msg)?;
+        }
+        log::debug!(target: "remp", "Polling incoming REMP messages delayer: {} forwarded, {} waiting", cnt, self.len().await);
+        Ok(cnt)
+    }
 }
 
 #[async_trait::async_trait]
@@ -329,6 +438,7 @@ pub struct RempManager {
 
     pub catchain_store: Arc<RempCatchainStore>,
     pub message_cache: Arc<MessageCache>,
+    incoming_delayer: RempDelayer,
     incoming_dispatcher: RempQueueDispatcher<RmqMessage, RempIncomingQueue>,
     pub collator_receipt_dispatcher: RempQueueDispatcher<CollatorResult, CollatorInterfaceWrapper>,
     pub response_sender: crossbeam_channel::Sender<(UInt256, Arc<RmqMessage>, RempMessageStatus)>
@@ -339,20 +449,24 @@ impl RempManager {
         -> (Self, RempInterfaceQueues) 
     {
         let (incoming_sender, incoming_receiver) = crossbeam_channel::unbounded();
+        let (delayed_incoming_sender, delayed_incoming_receiver) = crossbeam_channel::unbounded();
         let (response_sender, response_receiver) = crossbeam_channel::unbounded();
         let message_cache = Arc::new(MessageCache::with_metrics(
             #[cfg(feature = "telemetry")]
             engine.remp_core_telemetry().cache_size_metric()
         ));
 
+        let mut delay_random_rng = rand::thread_rng();
+        let delay_random_seed: u64 = delay_random_rng.gen();
         let collator_interface_wrapper = CollatorInterfaceWrapper::new(engine.clone());
         return (RempManager {
-            options: opt,
+            options: opt.clone(),
             catchain_store: Arc::new(RempCatchainStore::new()),
             message_cache: message_cache.clone(),
+            incoming_delayer: RempDelayer::new(delay_random_seed, &opt, incoming_receiver, delayed_incoming_sender),
             incoming_dispatcher: RempQueueDispatcher::with_metric(
                 "incoming".to_string(),
-                RempIncomingQueue::new(engine.clone(), incoming_receiver),
+                RempIncomingQueue::new(engine.clone(), delayed_incoming_receiver),
                 #[cfg(feature = "telemetry")]
                 engine.remp_core_telemetry().incoming_queue_size_metric(),
                 #[cfg(feature = "telemetry")]
@@ -390,6 +504,14 @@ impl RempManager {
     }
 
     pub async fn poll_incoming(&self, shard: &ShardIdent) -> (Option<Arc<RmqMessage>>, usize) {
+        match self.incoming_delayer.poll_incoming().await {
+            Ok(n) => {
+                log::trace!(target: "remp", "Polling REMP incoming delayer queue: {} messages processed", n);
+            },
+            Err(e) => {
+                log::error!(target: "remp", "Cannot poll REMP incoming delayer queue: {}", e);
+            }
+        }
         return self.incoming_dispatcher.poll(shard).await;
     }